# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "California Housing Market API",
    "version": "1.0.0"
  },
  "paths": {
    "/api/v1/meta": {
      "get": {
        "summary": "Meta",
        "operationId": "meta_api_v1_meta_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Meta"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/filters": {
      "get": {
        "summary": "Filters",
        "operationId": "filters_api_v1_filters_get",
        "parameters": [
          {
            "name": "version",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "uuid"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Version"
            }
          },
          {
            "name": "month",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "pattern": "^\\d{4}-(0[1-9]|1[0-2])$"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Month"
            }
          },
          {
            "name": "county",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "maxLength": 100
                },
                {
                  "type": "null"
                }
              ],
              "title": "County"
            }
          },
          {
            "name": "city",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "maxLength": 100
                },
                {
                  "type": "null"
                }
              ],
              "title": "City"
            }
          },
          {
            "name": "zip",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "pattern": "^\\d{5}$"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Zip"
            }
          },
          {
            "name": "subtype",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "maxLength": 100
                },
                {
                  "type": "null"
                }
              ],
              "title": "Subtype"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FilterOptions"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/summary": {
      "get": {
        "summary": "Summary",
        "operationId": "summary_api_v1_summary_get",
        "parameters": [
          {
            "name": "version",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "uuid"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Version"
            }
          },
          {
            "name": "month",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "pattern": "^\\d{4}-(0[1-9]|1[0-2])$"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Month"
            }
          },
          {
            "name": "county",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "maxLength": 100
                },
                {
                  "type": "null"
                }
              ],
              "title": "County"
            }
          },
          {
            "name": "city",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "maxLength": 100
                },
                {
                  "type": "null"
                }
              ],
              "title": "City"
            }
          },
          {
            "name": "zip",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "pattern": "^\\d{5}$"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Zip"
            }
          },
          {
            "name": "subtype",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "maxLength": 100
                },
                {
                  "type": "null"
                }
              ],
              "title": "Subtype"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Summary"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/trends": {
      "get": {
        "summary": "Trends",
        "operationId": "trends_api_v1_trends_get",
        "parameters": [
          {
            "name": "version",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "uuid"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Version"
            }
          },
          {
            "name": "month",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "pattern": "^\\d{4}-(0[1-9]|1[0-2])$"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Month"
            }
          },
          {
            "name": "county",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "maxLength": 100
                },
                {
                  "type": "null"
                }
              ],
              "title": "County"
            }
          },
          {
            "name": "city",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "maxLength": 100
                },
                {
                  "type": "null"
                }
              ],
              "title": "City"
            }
          },
          {
            "name": "zip",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "pattern": "^\\d{5}$"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Zip"
            }
          },
          {
            "name": "subtype",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "maxLength": 100
                },
                {
                  "type": "null"
                }
              ],
              "title": "Subtype"
            }
          },
          {
            "name": "period",
            "in": "query",
            "required": false,
            "schema": {
              "enum": [
                "24",
                "all"
              ],
              "type": "string",
              "default": "24",
              "title": "Period"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Trends"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/map": {
      "get": {
        "summary": "Map Data",
        "operationId": "map_data_api_v1_map_get",
        "parameters": [
          {
            "name": "version",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "uuid"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Version"
            }
          },
          {
            "name": "month",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "pattern": "^\\d{4}-(0[1-9]|1[0-2])$"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Month"
            }
          },
          {
            "name": "county",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "maxLength": 100
                },
                {
                  "type": "null"
                }
              ],
              "title": "County"
            }
          },
          {
            "name": "city",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "maxLength": 100
                },
                {
                  "type": "null"
                }
              ],
              "title": "City"
            }
          },
          {
            "name": "zip",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "pattern": "^\\d{5}$"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Zip"
            }
          },
          {
            "name": "subtype",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "maxLength": 100
                },
                {
                  "type": "null"
                }
              ],
              "title": "Subtype"
            }
          },
          {
            "name": "level",
            "in": "query",
            "required": false,
            "schema": {
              "enum": [
                "county",
                "zip"
              ],
              "type": "string",
              "default": "county",
              "title": "Level"
            }
          },
          {
            "name": "metric",
            "in": "query",
            "required": false,
            "schema": {
              "enum": [
                "median_price",
                "closed_sales",
                "concentration"
              ],
              "type": "string",
              "default": "median_price",
              "title": "Metric"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MapData"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/competitive": {
      "get": {
        "summary": "Competitive",
        "operationId": "competitive_api_v1_competitive_get",
        "parameters": [
          {
            "name": "version",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "uuid"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Version"
            }
          },
          {
            "name": "month",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "pattern": "^\\d{4}-(0[1-9]|1[0-2])$"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Month"
            }
          },
          {
            "name": "county",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "maxLength": 100
                },
                {
                  "type": "null"
                }
              ],
              "title": "County"
            }
          },
          {
            "name": "city",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "maxLength": 100
                },
                {
                  "type": "null"
                }
              ],
              "title": "City"
            }
          },
          {
            "name": "zip",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "pattern": "^\\d{5}$"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Zip"
            }
          },
          {
            "name": "subtype",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "maxLength": 100
                },
                {
                  "type": "null"
                }
              ],
              "title": "Subtype"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Competitive"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Area": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "value": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Value"
          },
          "closed_sales": {
            "type": "integer",
            "title": "Closed Sales"
          },
          "median_price": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Median Price"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "value",
          "closed_sales",
          "median_price"
        ],
        "title": "Area"
      },
      "Cell": {
        "properties": {
          "latitude": {
            "type": "number",
            "title": "Latitude"
          },
          "longitude": {
            "type": "number",
            "title": "Longitude"
          },
          "count": {
            "type": "integer",
            "title": "Count"
          }
        },
        "type": "object",
        "required": [
          "latitude",
          "longitude",
          "count"
        ],
        "title": "Cell"
      },
      "Competitive": {
        "properties": {
          "meta": {
            "$ref": "#/components/schemas/Meta"
          },
          "agents": {
            "items": {
              "$ref": "#/components/schemas/Entity"
            },
            "type": "array",
            "title": "Agents"
          },
          "offices": {
            "items": {
              "$ref": "#/components/schemas/Entity"
            },
            "type": "array",
            "title": "Offices"
          },
          "closed_sales": {
            "type": "integer",
            "title": "Closed Sales"
          },
          "sales_volume": {
            "type": "number",
            "title": "Sales Volume"
          }
        },
        "type": "object",
        "required": [
          "meta",
          "agents",
          "offices",
          "closed_sales",
          "sales_volume"
        ],
        "title": "Competitive"
      },
      "Entity": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "closed_sales": {
            "type": "integer",
            "title": "Closed Sales"
          },
          "sales_volume": {
            "type": "number",
            "title": "Sales Volume"
          },
          "median_price": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Median Price"
          },
          "market_share": {
            "type": "number",
            "title": "Market Share"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "closed_sales",
          "sales_volume",
          "median_price",
          "market_share"
        ],
        "title": "Entity"
      },
      "FilterOptions": {
        "properties": {
          "meta": {
            "$ref": "#/components/schemas/Meta"
          },
          "counties": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Counties"
          },
          "cities": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Cities"
          },
          "zips": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Zips"
          },
          "subtypes": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Subtypes"
          }
        },
        "type": "object",
        "required": [
          "meta",
          "counties",
          "cities",
          "zips",
          "subtypes"
        ],
        "title": "FilterOptions"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "MapData": {
        "properties": {
          "meta": {
            "$ref": "#/components/schemas/Meta"
          },
          "metric": {
            "type": "string",
            "enum": [
              "median_price",
              "closed_sales",
              "concentration"
            ],
            "title": "Metric"
          },
          "level": {
            "type": "string",
            "enum": [
              "county",
              "zip"
            ],
            "title": "Level"
          },
          "areas": {
            "items": {
              "$ref": "#/components/schemas/Area"
            },
            "type": "array",
            "title": "Areas"
          },
          "cells": {
            "items": {
              "$ref": "#/components/schemas/Cell"
            },
            "type": "array",
            "title": "Cells"
          },
          "unmapped_sales": {
            "type": "integer",
            "title": "Unmapped Sales"
          }
        },
        "type": "object",
        "required": [
          "meta",
          "metric",
          "level",
          "areas",
          "cells",
          "unmapped_sales"
        ],
        "title": "MapData"
      },
      "Meta": {
        "properties": {
          "version": {
            "type": "string",
            "title": "Version"
          },
          "start_month": {
            "type": "string",
            "title": "Start Month"
          },
          "end_month": {
            "type": "string",
            "title": "End Month"
          },
          "published_at": {
            "type": "string",
            "title": "Published At"
          },
          "demo": {
            "type": "boolean",
            "title": "Demo"
          },
          "months": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Months"
          },
          "source": {
            "type": "string",
            "title": "Source",
            "default": "CRMLS via Trestle"
          },
          "units": {
            "additionalProperties": {
              "type": "string"
            },
            "type": "object",
            "title": "Units"
          },
          "refresh_schedule": {
            "type": "string",
            "title": "Refresh Schedule",
            "default": "Monthly \u00b7 day 7 \u00b7 6:00 AM Pacific"
          }
        },
        "type": "object",
        "required": [
          "version",
          "start_month",
          "end_month",
          "published_at",
          "demo",
          "months"
        ],
        "title": "Meta"
      },
      "Metrics": {
        "properties": {
          "new_listings": {
            "type": "integer",
            "title": "New Listings",
            "default": 0
          },
          "closed_sales": {
            "type": "integer",
            "title": "Closed Sales",
            "default": 0
          },
          "median_price": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Median Price"
          },
          "days_on_market": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Days On Market"
          },
          "price_sqft": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Price Sqft"
          },
          "price_ratio": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Price Ratio"
          },
          "mortgage_rate": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Mortgage Rate"
          },
          "days_sample": {
            "type": "integer",
            "title": "Days Sample",
            "default": 0
          },
          "sqft_sample": {
            "type": "integer",
            "title": "Sqft Sample",
            "default": 0
          },
          "ratio_sample": {
            "type": "integer",
            "title": "Ratio Sample",
            "default": 0
          }
        },
        "type": "object",
        "title": "Metrics"
      },
      "Summary": {
        "properties": {
          "meta": {
            "$ref": "#/components/schemas/Meta"
          },
          "month": {
            "type": "string",
            "title": "Month"
          },
          "current": {
            "$ref": "#/components/schemas/Metrics"
          },
          "previous": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Metrics"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "meta",
          "month",
          "current",
          "previous"
        ],
        "title": "Summary"
      },
      "TrendPoint": {
        "properties": {
          "new_listings": {
            "type": "integer",
            "title": "New Listings",
            "default": 0
          },
          "closed_sales": {
            "type": "integer",
            "title": "Closed Sales",
            "default": 0
          },
          "median_price": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Median Price"
          },
          "days_on_market": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Days On Market"
          },
          "price_sqft": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Price Sqft"
          },
          "price_ratio": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Price Ratio"
          },
          "mortgage_rate": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Mortgage Rate"
          },
          "days_sample": {
            "type": "integer",
            "title": "Days Sample",
            "default": 0
          },
          "sqft_sample": {
            "type": "integer",
            "title": "Sqft Sample",
            "default": 0
          },
          "ratio_sample": {
            "type": "integer",
            "title": "Ratio Sample",
            "default": 0
          },
          "month": {
            "type": "string",
            "title": "Month"
          }
        },
        "type": "object",
        "required": [
          "month"
        ],
        "title": "TrendPoint"
      },
      "Trends": {
        "properties": {
          "meta": {
            "$ref": "#/components/schemas/Meta"
          },
          "points": {
            "items": {
              "$ref": "#/components/schemas/TrendPoint"
            },
            "type": "array",
            "title": "Points"
          }
        },
        "type": "object",
        "required": [
          "meta",
          "points"
        ],
        "title": "Trends"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}